Moved mail_api and mail_gmail_impl code to mail_client
