--- mail_api/pyproject.toml
@@ -1,11 +1,0 @@
-[project]
-name = "mail-api"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = ">=3.11"
-dependencies = []
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
