--- mail_gmail_impl/pyproject.toml
@@ -1,21 +1,0 @@
-[project]
-name = "mail_gmail_impl"
-version = "0.1.0"
-description = "Gmail implementation of mail_api interfaces"
-readme = "README.md"
-requires-python = ">=3.11"
-dependencies = [
-    "mail_api",
-    "google-api-python-client>=2.113.0",
-    "google-auth>=2.28.1",
-    "google-auth-oauthlib>=1.2.0",
-    "google-auth-httplib2>=0.1.1",
-    "html2text>=2023.0.0",
-]
-
-[build-system]
-requires = ["setuptools>=42", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[tool.uv.sources]
-mail_api = { workspace = true }
